backend/pyproject.toml
[tool.poetry]
name = "timeline"
version = "0.1.0"
description = "Timeline für Ereignisse, die im Leben von Maxine passieren."
authors = ["Kevin Bieri <bierike@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
Django = "^3.2.8"
psycopg2 = "^2.9.5"
djangorestframework = "^3.14.0"
django-extensions = "^3.2.1"
django-filter = "^22.1"
django-cors-headers = "^3.13.0"
django-editorjs-fields = "^0.2.7"
django-cleanup = "^6.0.0"
sorl-thumbnail = "^12.9.0"
django-heroku = "^0.3.1"
django-storages = "^1.13.2"
boto3 = "^1.26.54"
django-tus = "^0.5.0"
gunicorn = "^20.1.0"
whitenoise = "^6.3.0"
django-rest-knox = "^4.2.0"
pyvips = "^2.2.1"
pillow = "^10.0.0"
ruff = "^0.2.1"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
model-bakery = "^1.10.1"
pluck = "^0.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 125
